refactor: cria arquivo na pasta common para funções formatadoras de cpf e aniversário
