+++ src/features/acf/modules/AcfDashboardPage/modules/PanelSelector/modules/dashboards/modules/table/modules/diabetes/columns.ts
@@ -1,8 +1,9 @@
+import { cpfOrBirthdayFormatter } from "@/common/cpfOrBirthday";
 import {
     renderStatusTagCell,
 } from "@/helpers/lista-nominal/renderCell";
 import type { GridColDef } from "@mui/x-data-grid";
-import { dateRenderCell, cpfOrBirthdayFormatter  } from "./columns.formatter";
+import { dateRenderCell } from "./columns.formatter";
 import { iconDetailsMap } from "./iconDetailsMap";
 //TODO investigar como a conversao para string no dataGrid
